pyproject.toml
[tool.poetry]
name = "benchmarking_tool"
version = "0.1.0"
description = ""
authors = ["Colin Wilk <colin.wilk@tum.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
celery = {extras = ["redis"], version = "^5.4.0"}
sqlalchemy = "^2.0.30"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
schedule = "^1.2.2"
loguru = "^0.7.2"
tabulate = "^0.9.0"
humanize = "^4.9.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pylint = "^3.2.3"
black = "^24.4.2"
isort = "^5.13.2"
yamllint = "^1.35.1"
types-tabulate = "^0.9.0.20240106"
celery-types = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
